Add files via upload
The game so far is a map which has functional hex grid responses to inputs such as a click. To run the game so far all you need is to have the main.PY and then the "game" folder installed in the same location with the contents of "HexTileMap" and "HexTileMap". They are crucial to running the code.